Seperating Earlier test cases and creation of a new scenario
