+++ src/test/java/testCases/HotelsPageSelection/CheckHotelOption.java
@@ -1,28 +1,29 @@
-package testCases.HotelTab;
+package testCases.HotelsPageSelection;
 
 import org.testng.annotations.Test;
 
 import BaseClasses.BaseTestClass;
 import BaseClasses.PagesBaseClass;
 import pages.CabsPage;
+import pages.HotelsPage;
 import pages.LandingPage;
 
 
-public class GetListOfAdultOption extends BaseTestClass {
-
+public class CheckHotelOption extends BaseTestClass {
+	
 	LandingPage landingPage;
 	CabsPage cabsPage;
+	HotelsPage hotelsPage;
 	
 	
 	@Test
-	public void listOfAdultOption() {
-		logger = report.createTest("Print list of available adult option");
+	public void checkHotelOption() {
+		logger = report.createTest("Check Hotel Option");
 		PagesBaseClass objPagesBaseClass = new PagesBaseClass(driver, logger);
 		landingPage = objPagesBaseClass.OpenApplication();
 		landingPage.closeLoginPopUp();
 		hotelsPage=landingPage.getHotelsPage();
-		hotelsPage.ClickOnHotelMenu();	
-		hotelsPage.ClickOnRoomAndGuest();	
-		hotelsPage.getListOfNumbersForAdultPerson();
+		hotelsPage.CheckHotelMenuOption();	
 	}
+
 }
