+++ src/test/java/testCases/VerifyRoomsAndAdultsFromHotelsPage/GetListOfAdultOption.java
@@ -1,4 +1,4 @@
-package testCases.HotelTab;
+package testCases.VerifyRoomsAndAdultsFromHotelsPage;
 
 import org.testng.annotations.Test;
 
@@ -8,21 +8,21 @@ import pages.CabsPage;
 import pages.LandingPage;
 
 
-public class VerifyRoomDetailBox extends BaseTestClass {
+public class GetListOfAdultOption extends BaseTestClass {
 
 	LandingPage landingPage;
 	CabsPage cabsPage;
 	
 	
 	@Test
-	public void checkRoomDetailBox() {
-		logger = report.createTest("Verify Room and Guest box ");
+	public void listOfAdultOption() {
+		logger = report.createTest("Print list of available adult option");
 		PagesBaseClass objPagesBaseClass = new PagesBaseClass(driver, logger);
 		landingPage = objPagesBaseClass.OpenApplication();
 		landingPage.closeLoginPopUp();
 		hotelsPage=landingPage.getHotelsPage();
 		hotelsPage.ClickOnHotelMenu();	
 		hotelsPage.ClickOnRoomAndGuest();	
-		hotelsPage.verifyRoomDetailsPopUp();
+		hotelsPage.getListOfNumbersForAdultPerson();
 	}
 }
